build,cocoa: adjust entitlements for core, build phases
This change fixes App Store distribution.  It's based on the "Embed an
Externally Built Tool" section of the guide at [1].

One issue before was that the signing phase ran before the copy phase,
which explains why the signing was always out of sync.  That was just
a straight-up build (very, very stupid) bug on my end.

Secondly, various docs mention that the helper entitlements should
_only_ contain the "app sandbox" and "inherit" entitlements, but
that's not true: every entitlement that the helper needs must be
listed in the file.

Thirdly, it's important that the signed helper has an ad-hoc signature
so its "identity" should be "-".

Lastly, the build phase that copies the helper into the app must be a
"Copy Files" phase and "Code Sign on Copy" must be toggled for the
helper.  This replaces the ad-hoc signature above with the right
signature for the bundle.

[1]: https://developer.apple.com/documentation/xcode/embedding-a-helper-tool-in-a-sandboxed-app